purge.sh: don't fail if directory to purge does not exist

+++ backup/etcd/purge.sh
@@ -7,6 +7,10 @@ BACKUPDIR=${BACKUPDIR:-"/opt/backup/etcd"}
 KEEP=${KEEP:-"10"}
 
 for d in hot cold_backup; do
+    if [ ! -d ${BACKUPDIR}/${d} ]; then
+      continue
+    fi
+
     ls -d $BACKUPDIR/$d/*.etcd | head -n "-${KEEP}" | while read i; do
       [ ! -z "$i" ] && [ -d $i ] && [[ $i =~ .etcd ]] && rm -rf $i
     done
